replace sqlite table of records with registry of JSON files #25
