+++ manager/app.py
@@ -11,12 +11,10 @@ from manager.models import db, User
 from manager.commands import (
     index,
     inspect_schema,
-    load_schema,
-    save_records,
-    reset_records,
     reset_user_password,
     bulk_update,
     set_all_ids,
+    registry_grp,
     create_user,
 )
 load_dotenv()
@@ -37,7 +35,7 @@ app = Flask(__name__)
 app.config['SQLALCHEMY_DATABASE_URI'] = f'sqlite:///{PROJECT_DIR}/data.db'
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 db.init_app(app)
- 
+    
 with app.app_context():
     db.create_all()
 
@@ -54,13 +52,11 @@ def load_user(user_id):
 
 app.cli.add_command(index)
 app.cli.add_command(inspect_schema)
-app.cli.add_command(load_schema)
-app.cli.add_command(save_records)
-app.cli.add_command(reset_records)
 app.cli.add_command(create_user)
 app.cli.add_command(reset_user_password)
 app.cli.add_command(bulk_update)
 app.cli.add_command(set_all_ids)
+app.cli.add_command(registry_grp)
 
 app.config['DEBUG'] = True
